nav bar based on user role, changes in backend as well
