+++ client/package.json
@@ -6,9 +6,10 @@
     "axios": "^0.17.1",
     "history": "^4.6.3",
     "prop-types": "^15.5.10",
-    "react": "^16.2.0",
-    "react-dom": "^16.2.0",
-    "react-redux": "^5.0.6",
+    "react": "^16.6.3",
+    "react-dom": "^16.6.3",
+    "react-file-base64": "^1.0.3",
+    "react-redux": "^5.1.1",
     "react-router": "^4.1.2",
     "react-router-dom": "^4.1.2",
     "redux": "^3.7.2",
@@ -17,12 +18,11 @@
     "redux-thunk": "^2.2.0"
   },
   "devDependencies": {
-    "eslint": "^5.5.0",
     "eslint-config-airbnb": "^17.1.0",
     "eslint-plugin-import": "^2.14.0",
-    "eslint-plugin-jsx-a11y": "^6.1.1",
+    "eslint-plugin-jsx-a11y": "^6.1.2",
     "eslint-plugin-react": "^7.11.1",
-    "react-scripts": "^1.1.5"
+    "react-scripts": "^2.1.1"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -30,5 +30,11 @@
     "test": "react-scripts test --env=jsdom",
     "eject": "react-scripts eject",
     "restart": "pm2 restart all"
-  }
+  },
+  "browserslist": [
+    ">0.2%",
+    "not dead",
+    "not ie <= 11",
+    "not op_mini all"
+  ]
 }
