+++ client/src/components/postcreate.js
@@ -1,7 +1,7 @@
 import React, { PureComponent } from 'react';
 import { Field, reduxForm } from 'redux-form';
 import { connect } from 'react-redux';
-import Header from './header';
+import Header from './header/header';
 import * as actions from '../actions';
 
 class Postcreate extends PureComponent {
