+++ client/src/components/postings.js
@@ -2,7 +2,7 @@ import React, { PureComponent } from 'react';
 import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
 import * as actions from '../actions';
-import Header from './header';
+import Header from './header/header';
 
 
 class Postings extends PureComponent {
